% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_ctd_report_section}
\alias{add_ctd_report_section}
\title{Add CTD profile and time series figures to the report}
\usage{
add_ctd_report_section(
  doc,
  ctd_data_full,
  lims_data_full,
  chl_stats,
  fig_num,
  cleanup
)
}
\description{
Add CTD profile and time series figures to the report
}
\keyword{internal}
