% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_mosaic_section}
\alias{add_mosaic_section}
\title{Add mosaic section to the report}
\usage{
add_mosaic_section(
  doc,
  mosaics,
  hab_species,
  region_label,
  cleanup,
  taxa_lookup = NULL
)
}
\description{
Add mosaic section to the report
}
\keyword{internal}
