% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{base_sweden_map}
\alias{base_sweden_map}
\title{Create a base map of Swedish waters}
\usage{
base_sweden_map()
}
\value{
A ggplot object with coastline, coordinate system, and theme.
}
\description{
Create a base map of Swedish waters
}
\keyword{internal}
