% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{compute_ctd_chl_avg}
\alias{compute_ctd_chl_avg}
\title{Compute 0-20m depth-averaged chlorophyll from CTD data}
\usage{
compute_ctd_chl_avg(ctd_data)
}
\arguments{
\item{ctd_data}{Data frame from \code{read_cnv_folder()}.}
}
\value{
Data frame with station_short, latitude, longitude, chl_mean.
}
\description{
Compute 0-20m depth-averaged chlorophyll from CTD data
}
