% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_chl_map}
\alias{create_chl_map}
\title{Create a standalone chlorophyll map}
\usage{
create_chl_map(chl_summary, title = "Chlorophyll")
}
\arguments{
\item{chl_summary}{Data frame with columns: station_short, latitude,
longitude, chl_mean.}

\item{title}{Plot title string.}
}
\value{
A ggplot object.
}
\description{
Create a standalone chlorophyll map
}
