% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_heatmap.R
\name{create_stacked_bar}
\alias{create_stacked_bar}
\title{Create a stacked bar chart of relative biovolume}
\usage{
create_stacked_bar(wide_summary, taxa_lookup = NULL, n_top = 10, title = "")
}
\arguments{
\item{wide_summary}{Wide-format data from \code{create_wide_summary()}.}

\item{taxa_lookup}{Optional taxa lookup table with \code{HAB} column.}

\item{n_top}{Number of top taxa to show individually. Default 10.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
HAB species are marked with a red asterisk (*) in the legend.
}
