% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{empty_classifications}
\alias{empty_classifications}
\title{Empty classification data.frame with the standard columns}
\usage{
empty_classifications()
}
\value{
A zero-row data.frame as returned by
\code{read_h5_classifications()}.
}
\description{
Empty classification data.frame with the standard columns
}
\keyword{internal}
