% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_io.R
\name{import_preview_ui}
\alias{import_preview_ui}
\title{Body of the "Import Corrections" confirmation dialog}
\usage{
import_preview_ui(
  corrections,
  thresholds,
  file_name,
  n_current,
  n_current_thresholds
)
}
\arguments{
\item{corrections}{Correction rows from \code{split_corrections_import()}.}

\item{thresholds}{Result of \code{adjustments_from_import()}.}

\item{file_name}{Name of the imported file.}

\item{n_current}{Number of corrections in the current session.}

\item{n_current_thresholds}{Number of threshold adjustments in the current
session.}
}
\value{
A \code{shiny::tagList}.
}
\description{
Body of the "Import Corrections" confirmation dialog
}
\keyword{internal}
