% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_available}
\alias{llm_available}
\title{Check if LLM text generation is available}
\usage{
llm_available()
}
\value{
TRUE if OPENAI_API_KEY, GEMINI_API_KEY or ANTHROPIC_API_KEY is
set, FALSE otherwise.
}
\description{
Check if LLM text generation is available
}
