% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{load_standard_stations}
\alias{load_standard_stations}
\title{Load standard monitoring stations with regional assignments}
\usage{
load_standard_stations()
}
\value{
A data.frame with columns: station_name (character), region (character).
Region values are e.g. "The Skagerrak", "The Eastern Baltic", etc.
}
\description{
Load standard monitoring stations with regional assignments
}
