% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{load_station_mapper}
\alias{load_station_mapper}
\title{Load station name synonym mapper}
\usage{
load_station_mapper()
}
\value{
A data.frame with columns: synonym, statn.
}
\description{
Maps various raw station name synonyms to canonical algaware station names.
}
