% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ctd.R
\name{mod_ctd_server}
\alias{mod_ctd_server}
\title{CTD Module Server}
\usage{
mod_ctd_server(id, config, rv)
}
\arguments{
\item{id}{Module namespace ID.}

\item{config}{Reactive values with settings.}

\item{rv}{Reactive values for shared app state.}
}
\value{
NULL (side effects only).
}
\description{
CTD Module Server
}
