% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{read_cnv_folder}
\alias{read_cnv_folder}
\title{Parse all CNV files from a folder tree (AlgAware-matched, for chl map)}
\usage{
read_cnv_folder(cnv_folder, algaware_stations)
}
\arguments{
\item{cnv_folder}{Path to parent Cnv/ folder (searched recursively).}

\item{algaware_stations}{Data frame from \code{load_algaware_stations()}.}
}
\value{
A data.frame or NULL.
}
\description{
Parse all CNV files from a folder tree (AlgAware-matched, for chl map)
}
